Docs: Use LDoc markdown support, exclude files.
Use markdown support, which is bundled together with LDoc.

Also, add a macro for the client type, which will simplify parameter
definitions. With this we can write `@client` instead of `@tparam
client` in order to denote that the parameter needs to be a client
structure. This simplifies `awful.client` module.

This also excludes some of the `.lua` files from generating the
documentation, because some of the files do not contain useful
information or any public functions at all. This make the docs easier to
navigate.

Remove layouts from documentation

Signed-off-by: Ignas Anikevicius (gns_ank) <anikevicius@gmail.com>